// src/app/model/PostpaidToPrepaid/PostToPrepaidView.model.ts
export interface PostToPrepaidViewModel {
    customerNumber: string | null;
    customerName: string | null;
    locationCode: string | null;
    locationName: string | null;
    nidNumber: string | null;
    mobileNumber: string | null;
    powerUtility: string | null;
    maxPower: string | null;
    transferDate: string | null;
}