// src/app/model/PostpaidToPrepaid/prepaidList.model.ts
export interface PrepaidListModel {
    customerNumber: string;
    customerName: string;
    locationName: string;
    locationCode: string;
    nidNumber: string;
    mobileNumber: string;
    powerUtility: string;
    maxPower: string;
    transID: string;
    fatherName: string;
    customerType: string;
    transferDate: string;
    preToPostId: number;
}